Add CLI instructions for `generate-schema`
When starting with gql.data, I spent quite some time trying to understand how to retrieve the schema without leaking API tokens in `tsconfig.json`.

This is the solution I'm using in my projects. It proves to be quite effective and even simple, despite the convoluted usage of single and double quotes required to control when shell interpolates the variable.

It uses the `dotenv` command, which is pretty common to have in dev dependencies.